pyproject.toml
[project]
name = "aikame"
version = "0.1.0"
description = "Your local RAG guide with internet access."
authors = [
	{name = "AatirNadim",email = "aatir.nadim@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"click (>=8.1.8,<9.0.0)",
	"pypdf2 (>=3.0.1,<4.0.0)",
	"faiss-cpu (>=1.9.0.post1,<2.0.0)",
	"numpy (>=2.2.2,<3.0.0)",
	"sentence-transformers (>=3.4.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aikame="aikame:cli"
